chat-ui/inspector: single-source trace/replay adaptation via /trace subpath
Move the Tier-A-compatible trace/replay adaptation LOGIC into @mcpjam/chat-ui
(new `/trace` subpath): adaptTraceToUiMessages, snapshotsToTraceWidgetSnapshots,
buildToolRenderOverridesFromSnapshots, buildPersistedExecutionReplay + trace
types. The inspector's trace-viewer-adapter and persisted-execution-replay
become boundary shims that delegate to the package — killing the largest
remaining duplication (~1000-line adapter).

Per the agreed placeholder-preserving bridge:
- chat-ui stays host-facing/structural: the widget/CSP/capability placeholders
  are now `unknown` (opaque pass-through the renderer never interprets), so a
  host's real McpUiResourceCsp / McpUiResourcePermissions / OpenAiAppsCapabilities
  bridge in without a cast.
- TraceEnvelope (eval spans / browser artifacts) stays inspector-side with real
  eval types; the package adapter consumes a structural TraceInput subset.
- The inspector shims keep the real MCP-Apps SDK types on AdaptedTraceResult /
  PersistedExecutionReplay and bridge the package placeholders at the boundary,
  so the adapter's ~15 consumers + ViewsTab are unchanged.

Also folds in the /thread-helpers wording fix: it "avoids the renderer/markdown
graph", not "React-free" (getToolStateMeta returns lucide icon components).

Inspector resolves /trace from source via vite/vitest/tsconfig aliases (ordered
before the bare @mcpjam/chat-ui alias). No new deps / lockfile change.

Verified: chat-ui typecheck/test/build (/trace dist emitted); inspector
typecheck:client; evals + share-usage + live-trace + chat-v2 thread suites
(615 tests, incl. the 27-test trace-adapter suite via the shim).
