+++ chat-ui/tsup.config.ts
@@ -3,7 +3,7 @@ import { defineConfig } from "tsup";
 // Tier A: a single public entrypoint. The transcript renderer and its
 // supporting types/helpers are all re-exported from `src/index.ts`.
 export default defineConfig({
-  entry: ["src/index.ts", "src/thread-helpers.ts"],
+  entry: ["src/index.ts", "src/thread-helpers.ts", "src/trace.ts"],
   format: ["esm"],
   target: "es2022",
   platform: "browser",
