+++ mcpjam-inspector/client/src/components/chat-v2/thread/thread-helpers.ts
@@ -1,6 +1,7 @@
-// Single-sourced from @mcpjam/chat-ui. These pure part/tool shape helpers were
+// Single-sourced from @mcpjam/chat-ui. These part/tool shape helpers were
 // extracted into the package (Tier A); the inspector re-exports them here so the
 // many existing `@/components/chat-v2/thread/thread-helpers` import sites keep
 // working against one implementation (no drift). The `/thread-helpers` subpath
-// pulls only the pure helpers — not the package's React/markdown graph.
+// avoids the package's renderer/markdown graph (not React-free — getToolStateMeta
+// returns lucide icon components).
 export * from "@mcpjam/chat-ui/thread-helpers";
