+++ mcpjam-inspector/client/vite.config.ts
@@ -25,12 +25,14 @@ const sdkHostConfigInternalEntry = path.resolve(
 // chat-ui/dist until it is built. Resolve the package from source so the
 // inspector's dev/build/typecheck/test never depend on a chat-ui build.
 const chatUiEntry = path.resolve(rootDir, "../chat-ui/src/index.ts");
-// Focused subpath: the pure thread helpers (no React/markdown graph). Inspector
-// modules single-source these from the package via this entry.
+// Focused subpaths resolved from source. They avoid the package's
+// renderer/markdown component graph (not React-free — thread-helpers still
+// exposes lucide icon components via getToolStateMeta).
 const chatUiThreadHelpersEntry = path.resolve(
   rootDir,
   "../chat-ui/src/thread-helpers.ts",
 );
+const chatUiTraceEntry = path.resolve(rootDir, "../chat-ui/src/trace.ts");
 // Bypass stale Vite optimized deps for MCP SDK auth helpers by resolving
 // directly to the installed ESM entrypoints.
 const mcpSdkClientAuthEntry = path.resolve(
@@ -74,8 +76,9 @@ export default defineConfig(({ mode }) => {
         "@repo/assets": path.resolve(clientDir, "src/assets"),
         "@/shared": path.resolve(clientDir, "../shared"),
         "@": path.resolve(clientDir, "./src"),
-        // More specific subpath must precede the bare alias (first match wins).
+        // More specific subpaths must precede the bare alias (first match wins).
         "@mcpjam/chat-ui/thread-helpers": chatUiThreadHelpersEntry,
+        "@mcpjam/chat-ui/trace": chatUiTraceEntry,
         "@mcpjam/chat-ui": chatUiEntry,
         "@mcpjam/sdk/browser": sdkBrowserEntry,
         "@mcpjam/sdk/host-config/internal": sdkHostConfigInternalEntry,
